# src/Venture.IdentityManagement.Api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Venture.IdentityManagement",
    "description": "",
    "termsOfService": "https://example.com/terms",
    "contact": {
      "name": "",
      "email": ""
    },
    "license": {
      "name": "Use under MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "v1"
  },
  "paths": {
    "/api/Privilege/{privilegeId}": {
      "get": {
        "tags": [
          "Privilege"
        ],
        "summary": "Get Privilege by id.",
        "description": "Get Privilege by id.",
        "operationId": "getPrivilegeById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "privilegeId",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetPrivilegeByIdResponse"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Privilege"
        ],
        "summary": "Delete Privilege.",
        "description": "Delete Privilege.",
        "operationId": "removePrivilege",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "privilegeId",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/RemovePrivilegeResponse"
            }
          }
        }
      }
    },
    "/api/Privilege": {
      "get": {
        "tags": [
          "Privilege"
        ],
        "summary": "Get Privileges.",
        "description": "Get Privileges.",
        "operationId": "getPrivileges",
        "produces": [
          "application/json"
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetPrivilegesResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Privilege"
        ],
        "summary": "Create Privilege.",
        "description": "Create Privilege.",
        "operationId": "createPrivilege",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreatePrivilegeRequest"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/CreatePrivilegeResponse"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Privilege"
        ],
        "summary": "Update Privilege.",
        "description": "Update Privilege.",
        "operationId": "updatePrivilege",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdatePrivilegeRequest"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/UpdatePrivilegeResponse"
            }
          }
        }
      }
    },
    "/api/Privilege/page/{pageSize}/{index}": {
      "get": {
        "tags": [
          "Privilege"
        ],
        "summary": "Get Privilege Page.",
        "description": "Get Privilege Page.",
        "operationId": "getPrivilegesPage",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "pageSize",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "index",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetPrivilegesPageResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "CreatePrivilegeRequest": {
      "type": "object",
      "properties": {
        "privilege": {
          "$ref": "#/definitions/PrivilegeDto"
        }
      }
    },
    "CreatePrivilegeResponse": {
      "type": "object",
      "properties": {
        "privilege": {
          "$ref": "#/definitions/PrivilegeDto"
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
    "GetPrivilegeByIdResponse": {
      "type": "object",
      "properties": {
        "privilege": {
          "$ref": "#/definitions/PrivilegeDto"
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "GetPrivilegesPageResponse": {
      "type": "object",
      "properties": {
        "length": {
          "format": "int32",
          "type": "integer"
        },
        "entities": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/PrivilegeDto"
          }
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "GetPrivilegesResponse": {
      "type": "object",
      "properties": {
        "privileges": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/PrivilegeDto"
          }
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
    "PrivilegeDto": {
      "type": "object",
      "properties": {
        "privilegeId": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
    "ProblemDetails": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "status": {
          "format": "int32",
          "type": "integer"
        },
        "detail": {
          "type": "string"
        },
        "instance": {
          "type": "string"
        }
      },
      "additionalProperties": { }
    },
    "RemovePrivilegeResponse": {
      "type": "object",
      "properties": {
        "privilege": {
          "$ref": "#/definitions/PrivilegeDto"
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
    "UpdatePrivilegeRequest": {
      "type": "object",
      "properties": {
        "privilege": {
          "$ref": "#/definitions/PrivilegeDto"
        }
      }
    },
    "UpdatePrivilegeResponse": {
      "type": "object",
      "properties": {
        "privilege": {
          "$ref": "#/definitions/PrivilegeDto"
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
  }
}